Chore: seperate types into folder
