+++ src/hooks/apis/Dashboard/useRecnetTodosQuery.ts
@@ -4,7 +4,7 @@ import { AxiosError } from 'axios';
 import { GET } from '@/apis/services/httpMethod';
 import { API_ENDPOINTS } from '@/constants/ApiEndpoints';
 import { QUERY_KEYS } from '@/constants/QueryKeys';
-import { RecentTodosResponse } from '@/types/response';
+import { RecentTodosResponse } from '@/types/Dashboard';
 
 export const recentTodosOptions = (): UseQueryOptions<
   RecentTodosResponse,
