+++ src/hooks/apis/Goals/useDeleteGoalMutation.ts
@@ -4,7 +4,7 @@ import { DELETE } from '@/apis/services/httpMethod';
 import { API_ENDPOINTS } from '@/constants/ApiEndpoints';
 import { QUERY_KEYS } from '@/constants/QueryKeys';
 import { notify } from '@/store/useToastStore';
-import { DeleteGoalResponse } from '@/types/response';
+import { DeleteGoalResponse } from '@/types/Goals';
 
 export const useDeleteGoalMutation = () => {
   const queryClient = useQueryClient();
