+++ src/hooks/apis/Goals/useEditGoalTitleMutation.ts
@@ -4,11 +4,8 @@ import { PUT } from '@/apis/services/httpMethod';
 import { API_ENDPOINTS } from '@/constants/ApiEndpoints';
 import { QUERY_KEYS } from '@/constants/QueryKeys';
 import { notify } from '@/store/useToastStore';
-import {
-  BaseInfiniteQueryResponse,
-  EditGoalTitleResponse,
-  GoalsDetailResponse,
-} from '@/types/response';
+import { EditGoalTitleResponse, GoalsDetailResponse } from '@/types/Goals';
+import { BaseInfiniteQueryResponse } from '@/types/response';
 
 interface PostDataTypes {
   title: string;
